hotfix: completed edits of the reviewer

+++ src/utils/Api.js
@@ -64,16 +64,15 @@ export class Api {
     }).then(this._checkResponse);
   }
 
-  setLike(cardId) {
+  changeLike(cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      }).then(this._checkResponse);
+    }
     return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
       method: "PUT",
-      headers: this._headers,
-    }).then(this._checkResponse);
-  }
-
-  deleteLike(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: "DELETE",
       headers: this._headers,
     }).then(this._checkResponse);
   }
@@ -88,4 +87,3 @@ const api = new Api({
 });
 
 export default api;
-
